Fixed scraping and processing steps, adapted to poetry environment. Renamed scraper files Added log folder for scraping error Refactored processing step
