+++ scripts/process.py
@@ -1,0 +1,9 @@
+from phdfinder.processing import processor
+
+"""
+CLI entrypoint for processing scraped job data.
+To run: `poetry run process`
+"""
+
+def main():
+    processor.main()
