+++ scripts/scrape.py
@@ -1,0 +1,5 @@
+# scripts/scrape.py
+from phdfinder.scraping.runner import run_all_scrapers
+
+def main():
+    run_all_scrapers()
