+++ src/phdfinder/app/streamlit_app.py
@@ -1,10 +1,10 @@
 import sqlite3
 from pathlib import Path
 import streamlit as st
-from db.viewer import download_filtered_tables, fetch_data
-from db.viewer import get_columns as _get_columns
-from db.viewer import get_tables as _get_tables
-from db.viewer import save_labels
+from viewer import download_filtered_tables, fetch_data
+from viewer import get_columns as _get_columns
+from viewer import get_tables as _get_tables
+from viewer import save_labels
 
 from config import pages_meta
 
@@ -58,7 +58,7 @@ def render_ui(metadata: list, conn: sqlite3.Connection):
         The app is designed to ease you PhD search by gathering data about open European PhD positions into one place.
         All data is scraped directly from the primary source - pages of the universities. The data is oriented to PhD job
         positions in biomedical field. However, some universities post various vacancies in the single list. Therefore,
-        don't be surprised to find positions in linguistics, theology or arts. This will be fixed in further updates of the scrapers.
+        don't be surprised to find positions in linguistics, theology or arts. This will be fixed in further updates of the universities.
        
         ---
         """
@@ -185,7 +185,7 @@ def render_ui(metadata: list, conn: sqlite3.Connection):
 
 def main():
     BASE = Path(__file__).parent
-    db_path = BASE.parent / "data" / "phd_jobs_in_schengen.db"
+    db_path = BASE.parent / "data" / "phd_jobs_in_schengen.processing"
 
     metadata = pages_meta
     conn = get_db_connection(db_path)
