+++ src/phdfinder/scraping/scraping_utils.py
@@ -50,7 +50,7 @@ def get_page_through_selenium(page, ad_element, return_driver=False):
         page_content_local = driver.page_source
         return page_content_local
 
-def get_db_path(relative_path="../../db/phd_jobs_in_schengen.db"):
+def get_db_path(relative_path="../../processing/phd_jobs_in_schengen.processing"):
     # Get the file path of the calling script
     calling_script = inspect.stack()[1].filename  # Get the filename of the caller
     # Get the directory of the calling script
